Move cursor to the next result if search performed at a result
If a search is performed, and the cursor is currently at the start of a
result, move beyond that to the next result. The only time a user might
want to keep the cursor at the current location is if they're only
searching to determine the number of matches; that's an exceptional
scenario, and so the default search behaviour shouldn't cater to that.