Update buffer rendering to use syntect's HighlightIterator
The previous implementation used scribe's TokenIterator, re-implementing
the calculation of styles that HighlightIterator already provides.
Additionally, this implementation should be much easier to cache, as the
parse state isn't tied to buffer data (the TokenIterator type can only
be used with the buffer version/state from which it was generated).

Fixes #22.