Improve heuristic for indent detection during newline insertion
Previously, we'd only consult the current line when determining
indentation for a new line. If the current line is blank, we don't
indent the new line at all; this is the wrong choice if the current
line is in an indented block.

We now search backwards from the current line, looking for a non-blank
line, and base the new line's indentation off of the that.

Fixes #103.