Allocate lexeme mapper values
As soon as we print the first mapped &str-based lexeme, which points to
the LexemeMapper instance, we can no longer mutate the mapper (because
there's an outstanding read-only reference to it). That prevents us from
entering the next loop iteration, which mutates the mapper to generate
the next mapped lexeme.