Allocate strings in print_lexeme
Well, this is disappointing. Since calling graphemes(true) on the lexeme
yields slices, and we're storing *those* in the terminal buffer,
Cow::Owned variants are dropped, which violates memory safety. This is a
shit workaround until I can come up with something that satisfies the
owned use case (which is just mapped lexemes at this point).