Migrate render state caching test into presenter
This test has also been updated to better focus on its intended purpose,
which is to confirm that the buffer's render cache is being fetched and
passed to the BufferRenderer. We already verify that its contents are
updated *correctly* in the BufferRenderer test suite, so we can reduce
these tests to a simple check of the number of render cache entries
before/after the method is called.