Make TermionTerminal current_position optional
There are scenarios where we don't have an initial starting position,
namely when starting up for the first time or resuming after a suspend.
Using a default position at 0,0 causes problems for the print method,
which uses this to determine whether the cursor should be explicitly
moved. Setting the position to None in these scenarios avoids errors in
print locations.