Fall back to default theme if custom can't be found
Even if a custom theme is set, that doesn't mean it exists. This updates
the Presenter type to handle this scenario, falling back to the default
theme when the custom/specified one can't be found.

It's particularly important to try to avoid raising errors when building
the presenter, because amp's default error handling relies on it. As a
result, we end up dumping the error to stdout and terminating the
process. There are legitimate cases where we should do that; finding a
custom theme isn't one of those.

Fixes #149.