Make OS clipboard support optional.
This lets the clipboard support depend on a cargo feature flag, which is
enabled by default. This means no visible change for normal users.

Disabling OS clipboard support can be useful when using Amp on e.g. Linux servers,
which normally do not have `libxcb` installed. And pulling in half of the
X11 world just for a terminal-based editor is kind of ridiculously.

Signed-off-by: Christoph Heiss <me@christoph-heiss.me>