extract precedence logic to private `value` function
The `value` function computes the value associated with a given key
for a given filename and extension. This was function was liberally
to reduce the repeated logic in the code, and *slightly modifies
behavior*, though in a backwards-compatible way. Some settings were
only customizeable when declared at a global level in the config
file, which may now be specified at the extension/filename level.
This does not (yet) apply to some of the more complicated settings
like `line_length_guide`: see the diff for mor details.