group justify command as a single operation
Prior to this commit, Amp would treat justification as two seperate
operations, leading to unexpected undo-redo behavior: Amp would first
get rid of the all the text and then replace it with the original
version upon two undos. The expected behavior is to treat the
justification as atomic, and Scribe's provided start/end_operation_group
methods are used to achieve that for that.