add --help message with more complex argument parsing
This implements a simple argument parser that isolates 'direct'
arguments (paths, etc) from `--key value` options, and treats the
argument --help specially to display a usage message.

For instance, the invocation `amp --key value src` would launch amp in
the directory src with the option key set to value. Options are not
currently used.

This commit makes it impossible to pass a file beginning with `--` as a
CLI argument: it will be assumed to be an option.