present statusline by left/right elements
The previous logic to display the statusline did so by right aligning
only the last element. This isn't sufficient if more than one element is
to appear on the right (e.g., a cursor position indicator). So, this
commit replaces the logic to first render the right elements as
compactly as possible, then render the first left elements as compactly
as possible, but expand out the last left element to fill the remaining
space.

Behavior is identical; this is purely an API change.