Upgrade to mio v1.0 with batched event handling
This is a major version bump with several breaking changes. The most
notable is the move towards edge triggers instead of level triggers,
which Amp relied on. This change forces consumers to read all pending
data to ensure a subsequent poll yields an event. This plays out in the
Terminal and EventListener types, where events are now greedily consumed
and sent to the main application thread, which has also been redesigned
to process multiple events in a single render cycle.

Another breaking change is the shift towards exterior mutability in the
Poll and Signals types. These are used in the Terminal type's listen
method, and required mutex wrappers to maintain this type's writable
interface across threads (it's shared between the threaded event
listener for input and main thread's view/renderer for output).