Use non-flake based direnv shell
Since this flake is configured to defer to a standard nix build under the hood
(which allows supporting both build methods), it makes sense to mimic that
approach for the direnv shell environment. By using the standard "use nix"
approach, we load shell.nix, which is also usable by the nix-shell command.