Remove lint workflow job
Until we can filter these to the just the changes in a PR, these are an
annoyance. This is wouldn't be so bad if the lints didn't change over time, but
because they do, running the job against the entire codebase introduces new
issues unrelated to the PR, even if it was previously compliant.