Persistent modes (#281)
There is a growing number of features/ideas that rely on persistent mode states.
These changes result in instantiating and storing modes once at startup, and
then reusing those when switching between them. Modes have a new reset method
that is called when they are activated, which allows them to selectively clear
state that should not persist between transitions.